services/twin/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
requests==2.31.0
httpx==0.24.1
pyyaml==6.0.1
